Remove all govuk unused scripts
[MAPDEV324]
